isolated answers for axios req
